capital_pay_inventory/requirements.txt
# Smart-Shop Stock Inventory (CapitalPay) — pinned for reproducibility
dash==2.18.2
dash-mantine-components==0.14.4
dash-ag-grid==31.2.0
plotly==5.24.1
pandas==2.2.3
SQLAlchemy==2.0.36
psycopg[binary]==3.2.9
python-dotenv==1.0.1
Werkzeug==3.1.3
bcrypt==4.2.1
openpyxl==3.1.5
reportlab==4.2.5
python-dateutil==2.9.0.post0
dash-iconify==0.1.2
